[VersionControl] Moved nested classes to their own files.
