+++ main/src/addins/VersionControl/MonoDevelop.VersionControl/MonoDevelop.VersionControl.Views/EditorCompareWidgetBase.cs
@@ -41,11 +41,10 @@ using MonoDevelop.Core.Text;
 using MonoDevelop.Ide.Editor;
 using Microsoft.VisualStudio.Text;
 using MonoDevelop.Ide.TextEditing;
-using MonoDevelop.Components.AtkCocoaHelper;
 
 namespace MonoDevelop.VersionControl.Views
 {
-	public abstract class EditorCompareWidgetBase : Gtk.Bin
+	public abstract partial class EditorCompareWidgetBase : Gtk.Bin
 	{
 		internal protected VersionControlDocumentInfo info;
 
@@ -732,573 +731,6 @@ namespace MonoDevelop.VersionControl.Views
 				}
 			}
 		}
-
-		class MiddleArea : DrawingArea
-		{
-			EditorCompareWidgetBase widget;
-			MonoTextEditor fromEditor, toEditor;
-			bool useLeft;
-
-			public Adjustment VAdjustment { get { return fromEditor.VAdjustment; } }
-
-			IEnumerable<Hunk> Diff {
-				get {
-					return useLeft ? widget.LeftDiff : widget.RightDiff;
-				}
-			}
-
-			public MiddleArea (EditorCompareWidgetBase widget, MonoTextEditor fromEditor, MonoTextEditor toEditor, bool useLeft)
-			{
-				this.widget = widget;
-				this.Events |= EventMask.PointerMotionMask | EventMask.ButtonPressMask;
-				this.fromEditor = fromEditor;
-				this.toEditor = toEditor;
-				this.useLeft = useLeft;
-				this.toEditor.EditorOptionsChanged += HandleToEditorhandleEditorOptionsChanged;
-				Accessible.SetRole (AtkCocoa.Roles.AXGroup, GettextCatalog.GetString ("Revert changes margin"));
-			}
-			
-			protected override void OnDestroyed ()
-			{
-				this.toEditor.EditorOptionsChanged -= HandleToEditorhandleEditorOptionsChanged;
-				this.ClearAccessibleButtons ();
-				base.OnDestroyed ();
-			}
-			
-			void HandleToEditorhandleEditorOptionsChanged (object sender, EventArgs e)
-			{
-				QueueDraw ();
-			}
-
-			Hunk selectedHunk = Hunk.Empty;
-			protected override bool OnMotionNotifyEvent (EventMotion evnt)
-			{
-				bool hideButton = widget.MainEditor.Document.IsReadOnly;
-				Hunk selectedHunk = Hunk.Empty;
-				if (!hideButton) {
-					int delta = widget.MainEditor.Allocation.Y - Allocation.Y;
-					if (Diff != null) {
-						foreach (var hunk in Diff) {
-							double z1 = delta + fromEditor.LineToY (hunk.RemoveStart) - fromEditor.VAdjustment.Value;
-							double z2 = delta + fromEditor.LineToY (hunk.RemoveStart + hunk.Removed) - fromEditor.VAdjustment.Value;
-							if (z1 == z2)
-								z2 = z1 + 1;
-	
-							double y1 = delta + toEditor.LineToY (hunk.InsertStart) - toEditor.VAdjustment.Value;
-							double y2 = delta + toEditor.LineToY (hunk.InsertStart + hunk.Inserted) - toEditor.VAdjustment.Value;
-	
-							if (y1 == y2)
-								y2 = y1 + 1;
-							double x, y, w, h;
-							GetButtonPosition (hunk, y1, y2, z1, z2, out x, out y, out w, out h);
-	
-							if (evnt.X >= x && evnt.X < x + w && evnt.Y >= y && evnt.Y < y + h) {
-								selectedHunk = hunk;
-								TooltipText = GettextCatalog.GetString ("Revert this change");
-								QueueDrawArea ((int)x, (int)y, (int)w, (int)h);
-								break;
-							}
-						}
-					}
-				} else {
-					selectedHunk = Hunk.Empty;
-				}
-
-				if (selectedHunk.IsEmpty)
-					TooltipText = null;
-
-				if (this.selectedHunk != selectedHunk) {
-					this.selectedHunk = selectedHunk;
-					QueueDraw ();
-				}
-				return base.OnMotionNotifyEvent (evnt);
-			}
-
-			protected override bool OnButtonPressEvent (EventButton evnt)
-			{
-				if (!evnt.TriggersContextMenu () && evnt.Button == 1 && !selectedHunk.IsEmpty) {
-					PerformRevert (selectedHunk);
-					return true;
-				}
-				return base.OnButtonPressEvent (evnt);
-			}
-
-			void PerformRevert (Hunk hunk)
-			{
-				try {
-					widget.UndoChange (fromEditor, toEditor, hunk);
-				} catch (Exception e) {
-					LoggingService.LogInternalError ("Error while undoing widget change.", e);
-				}
-			}
-
-			protected override bool OnLeaveNotifyEvent (EventCrossing evnt)
-			{
-				selectedHunk = Hunk.Empty;
-				TooltipText = null;
-				QueueDraw ();
-				return base.OnLeaveNotifyEvent (evnt);
-			}
-
-			const int buttonSize = 16;
-
-			public bool GetButtonPosition (Hunk hunk, double y1, double y2, double z1, double z2, out double x, out double y, out double w, out double h)
-			{
-				if (hunk.Removed > 0) {
-					var b1 = z1;
-					var b2 = z2;
-					x = useLeft ? 0 : Allocation.Width - buttonSize;
-					y = b1;
-					w = buttonSize;
-					h = b2 - b1;
-					return hunk.Inserted > 0;
-				} else {
-					var b1 = y1;
-					var b2 = y2;
-
-					x = useLeft ? Allocation.Width - buttonSize : 0;
-					y = b1;
-					w = buttonSize;
-					h = b2 - b1;
-					return  hunk.Removed > 0;
-				}
-			}
-
-			void DrawArrow (Cairo.Context cr, double x, double y)
-			{
-				if (useLeft) {
-					cr.MoveTo (x - 2, y - 3);
-					cr.LineTo (x + 2, y);
-					cr.LineTo (x - 2, y + 3);
-				} else {
-					cr.MoveTo (x + 2, y - 3);
-					cr.LineTo (x - 2, y);
-					cr.LineTo (x + 2, y + 3);
-				}
-			}
-
-			static void DrawCross (Cairo.Context cr, double x, double y)
-			{
-				cr.MoveTo (x - 2, y - 3);
-				cr.LineTo (x + 2, y + 3);
-				cr.MoveTo (x + 2, y - 3);
-				cr.LineTo (x - 2, y + 3);
-			}
-
-			protected override bool OnExposeEvent (EventExpose evnt)
-			{
-				bool hideButton = widget.MainEditor.Document.IsReadOnly;
-				using (Cairo.Context cr = Gdk.CairoHelper.Create (evnt.Window)) {
-					cr.Rectangle (evnt.Region.Clipbox.X, evnt.Region.Clipbox.Y, evnt.Region.Clipbox.Width, evnt.Region.Clipbox.Height);
-					cr.Clip ();
-					int delta = widget.MainEditor.Allocation.Y - Allocation.Y;
-					if (Diff != null) {
-						foreach (Hunk hunk in Diff) {
-							double z1 = delta + fromEditor.LineToY (hunk.RemoveStart) - fromEditor.VAdjustment.Value;
-							double z2 = delta + fromEditor.LineToY (hunk.RemoveStart + hunk.Removed) - fromEditor.VAdjustment.Value;
-							if (z1 == z2)
-								z2 = z1 + 1;
-	
-							double y1 = delta + toEditor.LineToY (hunk.InsertStart) - toEditor.VAdjustment.Value;
-							double y2 = delta + toEditor.LineToY (hunk.InsertStart + hunk.Inserted) - toEditor.VAdjustment.Value;
-	
-							if (y1 == y2)
-								y2 = y1 + 1;
-	
-							if (!useLeft) {
-								var tmp = z1;
-								z1 = y1;
-								y1 = tmp;
-	
-								tmp = z2;
-								z2 = y2;
-								y2 = tmp;
-							}
-	
-							int x1 = 0;
-							int x2 = Allocation.Width;
-	
-							if (!hideButton) {
-								if (useLeft && hunk.Removed > 0 || !useLeft && hunk.Removed == 0) {
-									x1 += 16;
-								} else {
-									x2 -= 16;
-								}
-							}
-	
-							if (z1 == z2)
-								z2 = z1 + 1;
-	
-							cr.MoveTo (x1, z1);
-							
-							cr.CurveTo (x1 + (x2 - x1) / 4, z1,
-								x1 + (x2 - x1) * 3 / 4, y1,
-								x2, y1);
-	
-							cr.LineTo (x2, y2);
-							cr.CurveTo (x1 + (x2 - x1) * 3 / 4, y2,
-								x1 + (x2 - x1) / 4, z2,
-								x1, z2);
-							cr.ClosePath ();
-							cr.SetSourceColor (GetColor (hunk, this.useLeft, false, 1.0));
-							cr.Fill ();
-	
-							cr.SetSourceColor (GetColor (hunk, this.useLeft, true, 1.0));
-							cr.MoveTo (x1, z1);
-							cr.CurveTo (x1 + (x2 - x1) / 4, z1,
-								x1 + (x2 - x1) * 3 / 4, y1,
-								x2, y1);
-							cr.Stroke ();
-							
-							cr.MoveTo (x2, y2);
-							cr.CurveTo (x1 + (x2 - x1) * 3 / 4, y2,
-								x1 + (x2 - x1) / 4, z2,
-								x1, z2);
-							cr.Stroke ();
-	
-							if (!hideButton) {
-								bool isButtonSelected = hunk == selectedHunk;
-	
-								double x, y, w, h;
-								bool drawArrow = useLeft ? GetButtonPosition (hunk, y1, y2, z1, z2, out x, out y, out w, out h) :
-									GetButtonPosition (hunk, z1, z2, y1, y2, out x, out y, out w, out h);
-	
-								cr.Rectangle (x, y, w, h);
-								if (isButtonSelected) {
-									int mx, my;
-									GetPointer (out mx, out my);
-								//	mx -= (int)x;
-								//	my -= (int)y;
-									using (var gradient = new Cairo.RadialGradient (mx, my, h, mx, my, 2)) {
-										var color = (MonoDevelop.Components.HslColor)Style.Mid (StateType.Normal);
-										color.L *= 1.05;
-										gradient.AddColorStop (0, color);
-										color.L *= 1.07;
-										gradient.AddColorStop (1, color);
-										cr.SetSource (gradient);
-									}
-								} else {
-									cr.SetSourceColor ((MonoDevelop.Components.HslColor)Style.Mid (StateType.Normal));
-								}
-								cr.FillPreserve ();
-								
-								cr.SetSourceColor ((MonoDevelop.Components.HslColor)Style.Dark (StateType.Normal));
-								cr.Stroke ();
-								cr.LineWidth = 1;
-								cr.SetSourceColor (MonoDevelop.Ide.Gui.Styles.BaseForegroundColor.ToCairoColor ());
-								if (drawArrow) {
-									DrawArrow (cr, x + w / 1.5, y + h / 2);
-									DrawArrow (cr, x + w / 2.5, y + h / 2);
-								} else {
-									DrawCross (cr, x + w / 2 , y + (h) / 2);
-								}
-								cr.Stroke ();
-							}
-						}
-					}
-				}
-				UpdateAccessiblity ();
-
-				return true;
-			}
-
-			internal void Refresh ()
-			{
-				QueueDraw ();
-			}
-
-			int oldVAdjusment1 = -1;
-			int oldVAdjusment2 = -1;
-
-			void UpdateAccessiblity ()
-			{
-				if (oldVAdjusment1 == (int)fromEditor.VAdjustment.Value &&
-					oldVAdjusment2 == (int)toEditor.VAdjustment.Value)
-					return;
-				oldVAdjusment1 = (int)fromEditor.VAdjustment.Value;
-				oldVAdjusment2 = (int)toEditor.VAdjustment.Value;
-
-				ClearAccessibleButtons ();
-
-				bool hideButton = widget.MainEditor.Document.IsReadOnly;
-				int delta = widget.MainEditor.Allocation.Y - Allocation.Y;
-
-				if (Diff != null) {
-					foreach (Hunk hunk in Diff) {
-						double z1 = delta + fromEditor.LineToY (hunk.RemoveStart) - fromEditor.VAdjustment.Value;
-						double z2 = delta + fromEditor.LineToY (hunk.RemoveStart + hunk.Removed) - fromEditor.VAdjustment.Value;
-						if (z1 == z2)
-							z2 = z1 + 1;
-
-						double y1 = delta + toEditor.LineToY (hunk.InsertStart) - toEditor.VAdjustment.Value;
-						double y2 = delta + toEditor.LineToY (hunk.InsertStart + hunk.Inserted) - toEditor.VAdjustment.Value;
-
-						if (y1 == y2)
-							y2 = y1 + 1;
-
-						if (!useLeft) {
-							var tmp = z1;
-							z1 = y1;
-							y1 = tmp;
-
-							tmp = z2;
-							z2 = y2;
-							y2 = tmp;
-						}
-
-						int x1 = 0;
-						int x2 = Allocation.Width;
-
-						if (!hideButton) {
-							if (useLeft && hunk.Removed > 0 || !useLeft && hunk.Removed == 0) {
-								x1 += 16;
-							} else {
-								x2 -= 16;
-							}
-						}
-
-						if (z1 == z2)
-							z2 = z1 + 1;
-
-						if (!hideButton) {
-							bool isButtonSelected = hunk == selectedHunk;
-
-							double x, y, w, h;
-							bool drawArrow = useLeft ? GetButtonPosition (hunk, y1, y2, z1, z2, out x, out y, out w, out h) :
-								GetButtonPosition (hunk, z1, z2, y1, y2, out x, out y, out w, out h);
-
-							AddAccessibleProxyButton (hunk, x, y, w, h);
-						}
-					}
-				}
-				Accessible.SetAccessibleChildren (accessibleButtons.Select (a => a.Accessible).ToArray ());
-			}
-
-			void AddAccessibleProxyButton (Hunk hunk, double x, double y, double w, double h)
-			{
-				var button = new ButtonAccessible (this, hunk);
-				button.SetBounds ((int)x, (int)y, (int)w, (int)h);
-				accessibleButtons.Add (button);
-			}
-
-			void ClearAccessibleButtons ()
-			{
-				foreach (var button in accessibleButtons) {
-					button.Dispose ();
-				}
-				accessibleButtons.Clear ();
-				Accessible.SetAccessibleChildren (Array.Empty<AccessibilityElementProxy> ());
-			}
-
-			List<ButtonAccessible> accessibleButtons = new List<ButtonAccessible> ();
-
-			class ButtonAccessible : IDisposable
-			{
-				MiddleArea widget;
-				Hunk hunk;
-
-				public AccessibilityElementProxy Accessible { get; private set; }
-
-				public ButtonAccessible (MiddleArea widget, Hunk hunk)
-				{
-					this.widget = widget;
-					this.hunk = hunk;
-
-					Accessible = AccessibilityElementProxy.ButtonElementProxy ();
-					Accessible.GtkParent = widget;
-					Accessible.PerformPress += PerformPress;
-
-					Accessible.SetRole (AtkCocoa.Roles.AXButton, widget.useLeft ? GettextCatalog.GetString ("Revert left changes button") : GettextCatalog.GetString ("Revert right changes button"));
-
-					if (hunk.Inserted > 0) {
-						Accessible.Label = GettextCatalog.GetString ("Revert {0} inserted lines starting at {1}", hunk.Inserted, hunk.InsertStart);
-					} else if (hunk.Removed > 0) {
-						Accessible.Label = GettextCatalog.GetString ("Revert {0} removed lines starting at {1}", hunk.Removed, hunk.RemoveStart);
-					} else {
-						Accessible.Label = GettextCatalog.GetString ("Revert {0} replaced lines starting at {1}", hunk.Removed, hunk.InsertStart);
-					}
-				}
-
-				public void SetBounds (int x, int y, int w, int h)
-				{
-					Accessible.FrameInGtkParent = new Rectangle (x, y, w, h);
-
-					var cocoaY = widget.Allocation.Height - y - h;
-
-					Accessible.FrameInParent = new Rectangle (x, cocoaY, w, h);
-				}
-
-				void PerformPress (object sender, EventArgs e)
-				{
-					widget.PerformRevert (hunk);
-				}
-
-				public void Dispose ()
-				{
-					if (Accessible == null)
-						return;
-					Accessible.PerformPress -= PerformPress;
-					Accessible = null;
-					widget = null;
-				}
-			}
-		}
-
-		class DiffScrollbar : DrawingArea
-		{
-			MonoTextEditor editor;
-			EditorCompareWidgetBase widget;
-			bool useLeftDiff;
-			bool paintInsert;
-			Adjustment vAdjustment;
-			
-			public DiffScrollbar (EditorCompareWidgetBase widget, MonoTextEditor editor, bool useLeftDiff, bool paintInsert)
-			{
-				this.editor = editor;
-				this.useLeftDiff = useLeftDiff;
-				this.paintInsert = paintInsert;
-				this.widget = widget;
-				vAdjustment = widget.vAdjustment;
-				vAdjustment.ValueChanged += HandleValueChanged;
-				WidthRequest = 50;
-
-				Events |= EventMask.ButtonPressMask | EventMask.ButtonReleaseMask | EventMask.ButtonMotionMask;
-
-				Show ();
-			}
-
-			protected override void OnDestroyed ()
-			{
-				base.OnDestroyed ();
-				if (vAdjustment != null) {
-					vAdjustment.ValueChanged -= HandleValueChanged;
-					vAdjustment = null;
-				}
-			}
-
-			void HandleValueChanged (object sender, EventArgs e)
-			{
-				QueueDraw ();
-			}
-
-			public void MouseMove (double y)
-			{
-				var adj = widget.vAdjustment;
-				double position = (y / Allocation.Height) * adj.Upper - (double)adj.PageSize / 2;
-				position = Math.Max (0, Math.Min (position, adj.Upper - adj.PageSize));
-				widget.vAdjustment.Value = position;
-			}
-
-			protected override bool OnMotionNotifyEvent (EventMotion evnt)
-			{
-				if (button != 0)
-					MouseMove (evnt.Y);
-				return base.OnMotionNotifyEvent (evnt);
-			}
-
-			uint button;
-
-			protected override bool OnButtonPressEvent (EventButton evnt)
-			{
-				button |= evnt.Button;
-				MouseMove (evnt.Y);
-				return base.OnButtonPressEvent (evnt);
-			}
-
-			protected override bool OnButtonReleaseEvent (EventButton evnt)
-			{
-				button &= ~evnt.Button;
-				return base.OnButtonReleaseEvent (evnt);
-			}
-
-			protected override bool OnExposeEvent (Gdk.EventExpose e)
-			{
-				if (widget.LeftDiff == null)
-					return true;
-				var adj = widget.vAdjustment;
-				
-				var diff = useLeftDiff ? widget.LeftDiff : widget.RightDiff;
-				
-				using (Cairo.Context cr = Gdk.CairoHelper.Create (e.Window)) {
-					cr.LineWidth = 1;
-					double curY = 0;
-					
-					foreach (var hunk in diff) {
-						double y, count;
-						if (paintInsert) {
-							y = hunk.InsertStart / (double)editor.LineCount;
-							count = hunk.Inserted / (double)editor.LineCount;
-						} else {
-							y = hunk.RemoveStart / (double)editor.LineCount;
-							count = hunk.Removed / (double)editor.LineCount;
-						}
-						
-						double start  = y *  Allocation.Height;
-						FillGradient (cr, 0.5 + curY, start - curY);
-						
-						curY = start;
-						double height = Math.Max (cr.LineWidth, count * Allocation.Height);
-						cr.Rectangle (0.5, 0.5 + curY, Allocation.Width, height);
-						cr.SetSourceColor (GetColor (hunk, !paintInsert, false, 1.0));
-						cr.Fill ();
-						curY += height;
-					}
-					
-					FillGradient (cr, 0.5 + curY, Allocation.Height - curY);
-					
-					int barPadding = 3;
-					var allocH = Allocation.Height;
-					var adjUpper = adj.Upper;
-					var barY = allocH * adj.Value / adjUpper + barPadding;
-					var barH = allocH * (adj.PageSize / adjUpper) - barPadding - barPadding;
-					DrawBar (cr, barY, barH);
-					
-					cr.Rectangle (0.5, 0.5, Allocation.Width - 1, Allocation.Height - 1);
-					cr.SetSourceColor ((HslColor)Style.Dark (StateType.Normal));
-					cr.Stroke ();
-				}
-				return true;
-			}
-			
-			void FillGradient (Cairo.Context cr, double y, double h)
-			{
-				cr.Rectangle (0.5, y, Allocation.Width, h);
-
-				// FIXME: VV: Remove gradient features
-				using (var grad = new Cairo.LinearGradient (0, y, Allocation.Width, y)) {
-					var col = (HslColor)Style.Base (StateType.Normal);
-					col.L *= 0.95;
-					grad.AddColorStop (0, col);
-					grad.AddColorStop (0.7, (HslColor)Style.Base (StateType.Normal));
-					grad.AddColorStop (1, col);
-					cr.SetSource (grad);
-					cr.Fill ();
-				}
-			}
-			
-			void DrawBar (Cairo.Context cr, double y, double h)
-			{
-				int barPadding = 3;
-				int barWidth = Allocation.Width - barPadding - barPadding;
-				
-				MonoDevelop.Components.CairoExtensions.RoundedRectangle (cr, 
-					barPadding,
-					y,
-					barWidth,
-					h,
-					barWidth / 2);
-				
-				var color = Ide.Gui.Styles.BaseBackgroundColor;
-				color.Light = 0.5;
-				cr.SetSourceColor (color.WithAlpha (0.6).ToCairoColor ());
-				cr.Fill ();
-			}
-	
-			static void IncPos (Hunk h, ref int pos)
-			{
-				pos += System.Math.Max (h.Inserted, h.Removed);
-			}
-		}
 		
 		protected virtual void OnDiffChanged (EventArgs e)
 		{
@@ -1307,6 +739,4 @@ namespace MonoDevelop.VersionControl.Views
 		
 		public event EventHandler DiffChanged;
 	}
-
 }
-
