+++ main/src/addins/VersionControl/MonoDevelop.VersionControl/MonoDevelop.VersionControl.csproj
@@ -522,6 +522,8 @@
     <Compile Include="Gui\MonoDevelop.VersionControl.Views.LogWidget.cs" />
     <Compile Include="MonoDevelop.VersionControl\VersionControlDocumentController.cs" />
     <Compile Include="MonoDevelop.VersionControl\Instrumentation.cs" />
+    <Compile Include="MonoDevelop.VersionControl.Views\EditorCompareWidgetBase.MiddleArea.cs" />
+    <Compile Include="MonoDevelop.VersionControl.Views\EditorCompareWidgetBase.DiffScrollbar.cs" />
   </ItemGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="MonoDevelop.VersionControl.Git.Tests" />
